+++ src/Fitter/Minimizer/include/RootFactoryInterface.h
@@ -48,7 +48,7 @@ public:
   [[nodiscard]] const std::unique_ptr<ROOT::Math::Minimizer> &getMinimizer() const{ return _rootMinimizer_; }
 
   // core
-  double evalFit( const double* parArray_ );
+
   void saveMinimizerSettings(TDirectory* saveDir_) const;
 
 protected:
